Add Assets.isAdded(), don't add duplicate assets.
Previously, if an asset with the same path was added more than once the
onload handler passed to loadAll() would never fire.  This is because the
load_count variable is incremented only once per asset/path.  This
subsequently causes the overall load percentage to be miscalculated in
assetLoaded().  Because of this, the total load percentage can't ever
reach 100% and the onload handler is never called.

To fix this, we check whether an asset with the given path has already
been added when Assets.add() is called.  If it already exists in
src_list, we ignore it and log a warning via jaws.log.warn().

Additionally, users can manually check to see if an asset has been added
by calling the publicly accessible Assets.isAdded() method.

Signed-off-by: Atom Macumber <amacumber@eagleproductivity.com>