Allow arbitrary tile constructor for TMXMap.layerAsTileMap().
Instead of hard-coding jaws.Sprite() as the constructor for the objects
that are created and inserted into the resulting jaws.TileMap that comes
out of TMXMap.layerAsTileMap(), an optional constructor can be passed.
The given constructor will receive the same basic input object ({x, y,
width, height, image}) and the resulting object will gain a 'properties'
uh... property that contains any properties assigned to the given tile in
the map data.

Signed-off-by: suicidepills <atom.mac@gmail.com>