Bugfix: Gamepad stopped working in Chrome.
It appears that navigator.webkitGetGamepads may have been removed from
Chrome recently.  This caused the connectMethod to be set to 'event'
instead of 'poll' which threw off Gamepad detection in Chrome.  Now we
check for both webkitGetGamepads and (the more current) getGamepads
methods when assigning a value to the connectMethod variable.

Signed-off-by: suicidepills <atom.mac@gmail.com>